add del tests
