+++ mini-projects/00-jq-rs/src/maps/mod.rs
@@ -3,6 +3,6 @@ pub mod array_slice;
 pub mod array_index;
 pub mod object_identifier;
 pub mod add;
-pub mod del;
+pub mod del_key;
 pub mod length;
 pub mod maps;
