+++ mini-projects/00-jq-rs/src/parser.rs
@@ -4,7 +4,7 @@ use crate::maps::object_identifier::ObjectIdentifierMap;
 use crate::maps::array_slice::ArraySliceMap;
 use crate::maps::identity::IdentityMap;
 use crate::maps::add::AddMap;
-use crate::maps::del::DelMap;
+use crate::maps::del_key::DelMapKey;
 use crate::maps::length::LengthMap;
 use crate::maps::array_index::ArrayIndexMap;
 use anyhow::Result;
@@ -14,7 +14,7 @@ pub fn parse(input: String) -> Result<Vec<Box<dyn Map>>> {
         Box::new(ObjectIdentifierMap { key: "".to_string() }),
         Box::new(IdentityMap),
         Box::new(AddMap {}),
-        Box::new(DelMap { key: "".to_string() }),
+        Box::new(DelMapKey { key: "".to_string() }),
         Box::new(LengthMap {}),
         Box::new(ArraySliceMap { from: 0, to: 0 }),
         Box::new(ArrayIndexMap { index: 0 }),
